Made correction with error with add info to DB from page

+++ src/main/java/com/el/spring/daoRepository/RoleDao.java
@@ -3,6 +3,5 @@ package com.el.spring.daoRepository;
 import com.el.spring.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleDao extends JpaRepository<Role, Long>{
-
+public interface RoleDao extends JpaRepository<Role, Integer>{
 }
